feat: Initial support for Promise
JavaScript `Promise` without a return type is now supported.

Some changes to C# interop from JavaScript has been made. The old
and deprecated `BINDING.bind_static_method` has been removed as it does
not handle `int?` arguments correctly. Instead, the new
`getAssemblyExports` is used. However, the API is async so a static
constructor for `JsObject` is required to initialize the exports as soon
as possible.

The sample program has been updated to include a `Promise` example.